+++ app/src/main/java/com/example/gretchen/agrodata/data/DBHelperUsers.java
@@ -14,7 +14,7 @@ public class DBHelperUsers extends SQLiteOpenHelper {
     //version number to upgrade database version
     //each time if you Add, Edit table, you need to change the
     //version number.
-    private static final int DATABASE_VERSION = 1;
+    private static final int DATABASE_VERSION = 4;
 
     // Database Name
     private static final String DATABASE_NAME = "users.db";
@@ -23,6 +23,7 @@ public class DBHelperUsers extends SQLiteOpenHelper {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
     }
 
+    //Creates user table
     @Override
     public void onCreate(SQLiteDatabase db) {
         //All necessary tables you like to create will create here
@@ -42,10 +43,13 @@ public class DBHelperUsers extends SQLiteOpenHelper {
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
         // Drop older table if existed, all data will be gone!!!
-        db.execSQL("DROP TABLE IF EXISTS " + User.TABLE);
 
-        // Create tables again
-        onCreate(db);
+        if (newVersion > oldVersion) {
+            db.execSQL("DROP TABLE IF EXISTS " + User.TABLE);
+
+            // Create tables again
+            onCreate(db);
+        }
 
     }
